% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management_functions.R
\name{deg2rad}
\alias{deg2rad}
\title{deg2rad}
\usage{
deg2rad(degrees)
}
\arguments{
\item{degrees}{A numeric measurement in degrees}
}
\value{
A numeric measurement in radians
}
\description{
Convert angular measurements from degrees to radians
}
\examples{
deg2rad(30)
deg2rad(180)
deg2rad(360)
deg2rad(720)
}
