% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management_functions.R
\name{depths2thicknesses}
\alias{depths2thicknesses}
\title{Convert depths to unit thicknesses}
\usage{
depths2thicknesses(depths, values, decreasing = F)
}
\arguments{
\item{depths}{A vector of depth points, the same length as 'values'.}

\item{values}{A vector of lithology/unit values, the same length as 'depths'.}

\item{decreasing}{A boolean dictating whether to sort depths increasing or decreasing}
}
\value{
A data.frame object with thicknesses in the first column, lithology/unit values in the second column.
}
\description{
Convert depths to unit thicknesses
}
\examples{
depths2thicknesses(c(1, 2, 3, 4, 5), c(1, 1, 2, 2, 1))
depths2thicknesses(c(1010, 1020, 1030, 1040, 1050), c(1, 1, 2, 2, 1))
}
