% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_tools.R
\name{find_peaks}
\alias{find_peaks}
\title{Locate peaks}
\usage{
find_peaks(y, x = seq(1, length(y)), threshold = NULL)
}
\arguments{
\item{y}{Input vector of data in which to find peaks.}

\item{x}{(optional) locations of data observations.}

\item{threshold}{An optional value to be used as a cut-off for peaks.}
}
\value{
A data.frame object with peak locations (x) and peak values (y).
}
\description{
Locate peaks
}
\examples{
x <- seq(0, 2*pi, length=1000)
y <- sin(2*x)
find_peaks(y, x)
}
