% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_tools.R
\name{find_troughs}
\alias{find_troughs}
\title{Locate troughs}
\usage{
find_troughs(y, x = seq(1, length(y)), threshold = NULL)
}
\arguments{
\item{y}{Input vector of data in which to find troughs.}

\item{x}{(optional) locations of data observations.}

\item{threshold}{An optional value to be used as a cut-off for troughs.}
}
\value{
A data.frame object with trough locations (x) and troughs value (y).
}
\description{
Locate troughs
}
\examples{
x <- seq(0, 2*pi, 1000)
y <- sin(2*x)
find_troughs(y, x)
}
