% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseR_functions.R
\name{lith_plot}
\alias{lith_plot}
\title{Lithologic Stratigraphic Column using base R graphics}
\usage{
lith_plot(
  z,
  lith,
  widthmap,
  colormap,
  method = "thicknesses",
  start = 0,
  decreasing = F,
  ...
)
}
\arguments{
\item{z}{A vector of depth points or thicknesses, the same length as 'lith'.}

\item{lith}{A vector of lithology/unit codes, the same length as 'z'.}

\item{widthmap}{A numeric value between (0,1), 'proportional', 'equal',
or a named vector specifying which width between (0,1) to use for each
'lith' ID.}

\item{colormap}{A named vector of colors (as HEX codes, RGB values, or color
strings recognized by R), defining which color to use for each 'lith' ID.}

\item{method}{'thicknesses' if 'z' represents thicknesses, 'midpoint' if
you wish to interpolate transitions as the midpoint between consecutive
different lithologies, or any other string value (in which case, unit
thicknesses will be computed and the function will be called again
recursively with 'thicknesses' as the 'method').}

\item{start}{The starting depth to use if plotting with unit thicknesses.}

\item{decreasing}{A Boolean specifying whether to plot with increasing or
decreasing depth.}

\item{...}{...}
}
\description{
Lithologic Stratigraphic Column using base R graphics
}
\examples{
plot.new()
lith_plot(1:5, c(1, 2, 3, 3, 1), widthmap='proportional',
          colormap=c('1'='blue', '2'='red', '3'='yellow'),
          method='other')
plot.new()
lith_plot(c(1, 1, 2, 1), c(1, 2, 3, 1), widthmap='proportional',
          colormap=c('1'='blue', '2'='red', '3'='yellow'),
          method='thicknesses')
}
