% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_tools.R
\name{meshgrid}
\alias{meshgrid}
\title{Meshgrid}
\usage{
meshgrid(x, y)
}
\arguments{
\item{x}{Input vector of x-coordinates.}

\item{y}{Input vector of y-coordinates.}
}
\value{
A list of length 2 with a matrix (X) of x-coordinates and a matrix
(Y) of y-coordinates. Each row of X is a copy of input 'x'; each column of Y
is a copy of input 'y'.
}
\description{
returns 2D grid coordinates from input x and y vectors.
}
\examples{
x <- 1:10
y <- 1:10
meshgrid(y, x)
}
