% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management_functions.R
\name{rad2deg}
\alias{rad2deg}
\title{rad2deg}
\usage{
rad2deg(radians)
}
\arguments{
\item{radians}{A numeric measurement in radians}
}
\value{
A numeric measurement in degrees
}
\description{
Convert angular measurements from radians to degrees
}
\examples{
deg2rad(pi/2)
deg2rad(pi)
deg2rad(2*pi)
deg2rad(6*pi)
}
