% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseR_functions.R
\name{ternary_plot}
\alias{ternary_plot}
\title{Ternary Plot}
\usage{
ternary_plot(a, b, c = NULL, increment = NULL, new_plot = T, ...)
}
\arguments{
\item{a}{Input vector of a-values in decimal form (a + b + c = 1.0)}

\item{b}{Input vector of b-values in decimal form (a + b + c = 1.0)}

\item{increment}{Decimal value specifying increment for minor grid lines.}

\item{new_plot}{Boolean value specifying whether to create a new plot object
or plot on an existing plot object.}

\item{...}{Remaining args are passed to points() function for plotting data
points.}
}
\description{
Plots ternary coordinates on a triangle plot
}
\examples{
a <- c(0.1, 0.5, 0.2)
b <- c(0.6, 0.35, 0.2)
ternary_plot(a, b, increment=0.1, col="blue", pch=19)
}
