% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management_functions.R
\name{thicknesses2depths}
\alias{thicknesses2depths}
\title{Convert unit thicknesses to depths}
\usage{
thicknesses2depths(thicknesses, values, start = 0)
}
\arguments{
\item{thicknesses}{A vector of unit thicknesses, the same length as 'values'.}

\item{values}{A vector of lithology/unit values, the same length as 'thicknesses'.}

\item{start}{A starting depth to which thicknesses will be added to produce depths}
}
\value{
A data.frame object with depths in the first column, lithology/unit values in the second column.
}
\description{
Convert unit thicknesses to depths
}
\examples{
thicknesses2depths(c(2, 2, 5, 1), c(1, 2, 3, 1), 0)
thicknesses2depths(c(10, 5, 2, 3), c(3, 1, 2, 3), 150)
}
